Refactor Poses to Own Folder
So they don't sit at the top level of the commands.
